Added multi step form for event, expanded event details

+++ app/Http/Controllers/EventController.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace App\Http\Controllers;
 
 use App\Http\Requests\EventRequest;
+use App\Models\Category;
 use App\Models\Event;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Storage;
@@ -16,6 +17,7 @@ class EventController extends Controller
     public function index(): Response
     {
         $user = auth()->user();
+
         return Inertia::render('Events/Index', [
             'events' => $user->events()->paginate(15),
         ]);
@@ -23,7 +25,11 @@ class EventController extends Controller
 
     public function create(): Response
     {
-        return Inertia::render('Events/Create');
+
+        return Inertia::render('Events/Create',
+            [
+                'categories' => Category::all(),
+            ]);
     }
 
     public function edit(Event $event): Response
@@ -39,9 +45,10 @@ class EventController extends Controller
             'event' => $event,
         ]);
     }
+
     public function store(EventRequest $request): RedirectResponse
     {
-        $event = Event::create($request->safe()->except(['image']));
+        $event = auth()->user()->events()->create($request->safe()->except(['image']));
 
         if ($request->hasFile('image')) {
             $path = $request->file('image')->store('events', 'public');
@@ -55,6 +62,9 @@ class EventController extends Controller
 
     public function update(EventRequest $request, Event $event): RedirectResponse
     {
+        if($request->user()->cannot('update', $event)) {
+            abort(403);
+        }
         $event->update($request->safe()->except('image'));
         if ($request->hasFile('image')) {
             if ($event->image) {
@@ -69,6 +79,9 @@ class EventController extends Controller
 
     public function destroy(Event $event): RedirectResponse
     {
+        if(auth()->user()->cannot('delete', $event)) {
+            abort(403);
+        }
         $event->delete();
 
         session()->flash('notification', [
